refactor(openomni): merge stakes fragment modules into stakes-contract
Fold stakes-policy.ts (4 LOC of policy constants), stakes-digest.ts
(7 LOC hashStakesValue), and stakes-schema-primitives.ts (9 LOC zod
primitives) into stakes-contract.ts. All three were cluster-internal
single-purpose fragments. Importers rewired; the contract/reference
import cycle is function-only (no load-time evaluation) and
noImportCycles is off in the shared biome core.
