--- packages/openomni/src/ledger/stakes-policy.ts
@@ -1,4 +1,0 @@
-export const STAKES_POLICY_VERSION = "stakes-policy-v1";
-export const STAKES_THETA = 1_000;
-export const STAKES_EPSILON = 1;
-export const STAKES_AMOUNT_DENOMINATION = "owner_base_budget_micro_unit";
